+++ src/main/java/fr/inrae/act/bagap/chloe/window/analysis/grid/HugeGridLandscapeMetricAnalysis.java
@@ -40,7 +40,7 @@ public class HugeGridLandscapeMetricAnalysis extends GridLandscapeMetricAnalysis
 		}
 		
 		// gestion des sorties
-		kernel().setOutDatas(new double[gridWidth][nbValues()]);
+		kernel().setOutDatas(new float[gridWidth][nbValues()]);
 		
 		int yGrid, tYs, localBufferROIYMin, localBufferROIYMax;
 		Rectangle roi;
