+++ src/main/java/fr/inrae/act/bagap/chloe/window/counting/BasicCounting.java
@@ -2,12 +2,12 @@ package fr.inrae.act.bagap.chloe.window.counting;
 
 public class BasicCounting extends Counting implements BasicCountingInterface {
 
-	public BasicCounting(double theoreticalSize) {
+	public BasicCounting(float theoreticalSize) {
 		super(theoreticalSize);
 	}
 
 	@Override
-	public void doSetCounts(double[] counts) {
+	public void doSetCounts(float[] counts) {
 		// do nothing special
 	}
